test(runtime): assign reset regression to llama
Keep execution-context reuse coverage inside an isolated MODEL.toml ownership boundary so model impact selection can schedule it.
